requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
pydantic-settings
argon2-cffi
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
google-genai
python-multipart
apscheduler
reportlab
